+++ app/controllers/challah/test/restrictions_controller.rb
@@ -3,16 +3,11 @@
 class Challah::Test::RestrictionsController < ApplicationController
   signin_required :only => [ :blah ]
   before_filter :signin_required, :only => [ :edit ]
-  restrict_to_permission :special, :only => [ :new ]
   restrict_to_authenticated :only => [ :show ]
 
   def index
     current_user
 
-    head :ok
-  end
-
-  def new
     head :ok
   end