+++ app/models/permission_role.rb
@@ -1,3 +1,0 @@
-class PermissionRole < ActiveRecord::Base
-  authable_permission_role
-end