+++ test/restrictions_controller_test.rb
@@ -19,11 +19,6 @@ class RestrictionsControllerTest < ActionController::TestCase
         get :edit
         assert_redirected_to '/sign-in'
       end
-
-      should "not get to the new page" do
-        get :new
-        assert_redirected_to '/sign-in'
-      end
     end
 
     context "with a regular user" do
@@ -45,49 +40,6 @@ class RestrictionsControllerTest < ActionController::TestCase
 
       should "get to the show page" do
         get :show
-        assert_response :success
-      end
-
-      should "not get to the new page" do
-        get :new
-
-        assert_template 'sessions/access_denied'
-        assert_response :unauthorized
-      end
-    end
-
-    context "with an admin user" do
-      setup do
-        @user = create(:admin_user)
-        @permission = create(:permission, :key => 'special')
-
-        signin_as(@user)
-      end
-
-      should "get to the index page" do
-        get :index
-        assert_response :success
-        assert_equal @user, assigns(:current_user)
-      end
-
-      should "get to the blah page" do
-        get :blah
-        assert_response :success
-        assert_equal @user, assigns(:current_user)
-      end
-
-      should "get to the edit page" do
-        get :edit
-        assert_response :success
-      end
-
-      should "get to the show page" do
-        get :show
-        assert_response :success
-      end
-
-      should "get to the new page" do
-        get :new
         assert_response :success
       end
     end
@@ -117,13 +69,6 @@ class RestrictionsControllerTest < ActionController::TestCase
           get :show, :key => @user.api_key
           assert_response :success
         end
-
-        should "not get to the new page" do
-          get :new, :key => @user.api_key
-
-          assert_template 'sessions/access_denied'
-          assert_response :unauthorized
-        end
       end
 
       context "and api_key functionality disabled" do
@@ -144,11 +89,6 @@ class RestrictionsControllerTest < ActionController::TestCase
 
         should "not get to the show page" do
           get :show, :key => @user.api_key
-          assert_redirected_to '/sign-in'
-        end
-
-        should "not get to the new page" do
-          get :new, :key => @user.api_key
           assert_redirected_to '/sign-in'
         end
       end