+++ app/models/permission_user.rb
@@ -1,3 +1,0 @@
-class PermissionUser < ActiveRecord::Base
-  authable_permission_user
-end